requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# LangChain and AI components - using more compatible versions
langchain==0.1.0
langchain-openai==0.0.2
openai==1.3.7

# Google Calendar API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Streamlit for frontend
streamlit==1.28.2

# HTTP requests
requests==2.31.0

# Date and time handling
python-dateutil==2.8.2

# Environment variables
python-dotenv==1.0.0

# CORS support
python-multipart==0.0.6

# Additional dependencies for better compatibility
numpy>=1.24.0
typing-extensions>=4.0.0

# Cloud deployment dependencies
gunicorn>=21.2.0 